[feat] (all) Add getTransactionInfo to utils interface
