+++ packages/ripple-payments/src/BaseRipplePayments.ts
@@ -189,61 +189,6 @@ export abstract class BaseRipplePayments<Config extends BaseRipplePaymentsConfig
     return null
   }
 
-  async getTransactionInfo(txId: string): Promise<RippleTransactionInfo> {
-    let tx
-    try {
-      tx = await this._retryDced(() => this.api.getTransaction(txId))
-    } catch (e) {
-      const eString = e.toString()
-      if (NOT_FOUND_ERRORS.some(type => eString.includes(type))) {
-        throw new Error(`Transaction not found: ${eString}`)
-      }
-      throw e
-    }
-    this.logger.debug('getTransaction', txId, tx)
-    if (tx.type !== 'payment') {
-      throw new Error(`Unsupported ripple tx type ${tx.type}`)
-    }
-    const { specification, outcome } = tx as FormattedPaymentTransaction
-    const { source, destination } = specification
-    const amountObject = ((source as any).maxAmount || source.amount) as Amount
-    if (amountObject.currency !== 'XRP') {
-      throw new Error(`Unsupported ripple tx currency ${amountObject.currency}`)
-    }
-    const fromIndex = this.resolveIndexFromAdjustment(source)
-    const toIndex = this.resolveIndexFromAdjustment(destination)
-    const amount = amountObject.value
-    const isSuccessful = outcome.result.startsWith('tes')
-    const isCostDestroyed = outcome.result.startsWith('tec')
-    const status = isSuccessful || isCostDestroyed ? TransactionStatus.Confirmed : TransactionStatus.Failed
-    const isExecuted = isSuccessful
-    const confirmationNumber = outcome.ledgerVersion
-    const ledger = await this._retryDced(() => this.api.getLedger({ ledgerVersion: confirmationNumber }))
-    const currentLedgerVersion = await this.getCurrentBlockNumber()
-    const confirmationId = ledger.ledgerHash
-    const confirmationTimestamp = outcome.timestamp ? new Date(outcome.timestamp) : null
-    return {
-      status,
-      id: tx.id,
-      fromIndex,
-      fromAddress: source.address,
-      fromExtraId: typeof source.tag !== 'undefined' ? String(source.tag) : null,
-      toIndex,
-      toAddress: destination.address,
-      toExtraId: typeof destination.tag !== 'undefined' ? String(destination.tag) : null,
-      amount: amount,
-      fee: outcome.fee,
-      sequenceNumber: String(tx.sequence),
-      confirmationId,
-      confirmationNumber: String(confirmationNumber),
-      confirmationTimestamp,
-      isExecuted,
-      isConfirmed: Boolean(confirmationNumber),
-      confirmations: currentLedgerVersion - confirmationNumber,
-      data: tx,
-    }
-  }
-
   async resolveFeeOption(feeOption: FeeOption): Promise<ResolvedFeeOption> {
     let targetFeeLevel: FeeLevel
     let targetFeeRate: string
