+++ packages/stellar-payments/src/BaseStellarPayments.ts
@@ -181,65 +181,6 @@ export abstract class BaseStellarPayments<Config extends BaseStellarPaymentsConf
     return null
   }
 
-  async getLatestBlock(): Promise<Stellar.ServerApi.LedgerRecord> {
-    const page = await this._retryDced(() => this.getApi().ledgers()
-      .order('desc')
-      .limit(1)
-      .call())
-    if (!page.records) {
-      throw new Error('Failed to get stellar ledger records')
-    }
-    return page.records[0]
-  }
-
-  async getTransactionInfo(txId: string): Promise<StellarTransactionInfo> {
-    let tx: Stellar.ServerApi.TransactionRecord
-    try {
-      tx = await this._retryDced(() => this.getApi().transactions().transaction(txId).call())
-    } catch (e) {
-      const eString = e.toString()
-      if (NOT_FOUND_ERRORS.some(type => eString.includes(type))) {
-        throw new Error(`Transaction not found: ${eString}`)
-      }
-      throw e
-    }
-    // this.logger.debug('getTransactionInfo', txId, omitHidden(tx))
-    const { amount, fee, fromAddress, toAddress } = await this._normalizeTxOperation(tx)
-    const fromIndex = this.resolveIndexFromAddressAndMemo(fromAddress, tx.memo)
-    const toIndex = this.resolveIndexFromAddressAndMemo(toAddress, tx.memo)
-    const confirmationNumber = tx.ledger_attr
-    const ledger = await this._retryDced(() => tx.ledger())
-    const currentLedger = await this.getLatestBlock()
-    const currentLedgerSequence = currentLedger.sequence
-    const confirmationId = ledger.hash
-    const confirmationTimestamp = ledger.closed_at ? new Date(ledger.closed_at) : null
-    const confirmations = currentLedgerSequence - confirmationNumber
-    const sequenceNumber = tx.source_account_sequence
-    const isExecuted = (tx as any).successful
-    const isConfirmed = Boolean(confirmationNumber)
-    const status = isConfirmed || isExecuted ? TransactionStatus.Confirmed : TransactionStatus.Pending
-    return {
-      status,
-      id: tx.id,
-      fromIndex,
-      fromAddress,
-      fromExtraId: null,
-      toIndex,
-      toAddress,
-      toExtraId: tx.memo || null,
-      amount: amount.toString(),
-      fee: fee.toString(),
-      sequenceNumber,
-      confirmationId,
-      confirmationNumber: String(confirmationNumber),
-      confirmationTimestamp,
-      isExecuted,
-      isConfirmed,
-      confirmations,
-      data: tx,
-    }
-  }
-
   async resolveFeeOption(feeOption: FeeOption): Promise<ResolvedFeeOption> {
     let targetFeeLevel: FeeLevel
     let targetFeeRate: string
