+++ packages/stellar-payments/src/StellarPaymentsUtil.ts
@@ -1,5 +1,6 @@
-import { AutoFeeLevels, BalanceResult, FeeLevel, FeeRate, FeeRateType, PaymentsUtils, Payport } from '@faast/payments-common'
+import { AutoFeeLevels, BalanceResult, FeeLevel, FeeRate, FeeRateType, PaymentsUtils, Payport, TransactionStatus } from '@faast/payments-common'
 import { isNil, assertType, Numeric, isMatchingError } from '@faast/ts-common'
+import * as Stellar from 'stellar-sdk'
 
 import {
   toMainDenominationString,
@@ -10,6 +11,7 @@ import {
 import { StellarConnected } from './StellarConnected'
 import { COIN_NAME, COIN_SYMBOL, DECIMAL_PLACES, MIN_BALANCE, NOT_FOUND_ERRORS } from './constants'
 import BigNumber from 'bignumber.js'
+import { StellarTransactionInfo } from './types'
 
 export class StellarPaymentsUtils extends StellarConnected implements PaymentsUtils {
 
@@ -151,6 +153,63 @@ export class StellarPaymentsUtils extends StellarConnected implements PaymentsUt
   async getAddressNextSequenceNumber(address: string) {
     const accountInfo = await this.loadAccountOrThrow(address)
     return new BigNumber(accountInfo.sequence).plus(1).toString()
+  }
+
+  async getLatestBlock(): Promise<Stellar.ServerApi.LedgerRecord> {
+    const page = await this._retryDced(() => this.getApi().ledgers()
+      .order('desc')
+      .limit(1)
+      .call())
+    if (!page.records) {
+      throw new Error('Failed to get stellar ledger records')
+    }
+    return page.records[0]
+  }
+
+  async getTransactionInfo(txId: string): Promise<StellarTransactionInfo> {
+    let tx: Stellar.ServerApi.TransactionRecord
+    try {
+      tx = await this._retryDced(() => this.getApi().transactions().transaction(txId).call())
+    } catch (e) {
+      const eString = e.toString()
+      if (NOT_FOUND_ERRORS.some(type => eString.includes(type))) {
+        throw new Error(`Transaction not found: ${eString}`)
+      }
+      throw e
+    }
+    // this.logger.debug('getTransactionInfo', txId, omitHidden(tx))
+    const { amount, fee, fromAddress, toAddress } = await this._normalizeTxOperation(tx)
+    const confirmationNumber = tx.ledger_attr
+    const ledger = await this._retryDced(() => tx.ledger())
+    const currentLedger = await this.getLatestBlock()
+    const currentLedgerSequence = currentLedger.sequence
+    const confirmationId = ledger.hash
+    const confirmationTimestamp = ledger.closed_at ? new Date(ledger.closed_at) : null
+    const confirmations = currentLedgerSequence - confirmationNumber
+    const sequenceNumber = tx.source_account_sequence
+    const isExecuted = (tx as any).successful
+    const isConfirmed = Boolean(confirmationNumber)
+    const status = isConfirmed || isExecuted ? TransactionStatus.Confirmed : TransactionStatus.Pending
+    return {
+      status,
+      id: tx.id,
+      fromIndex: null,
+      fromAddress,
+      fromExtraId: null,
+      toIndex: null,
+      toAddress,
+      toExtraId: tx.memo || null,
+      amount: amount.toString(),
+      fee: fee.toString(),
+      sequenceNumber,
+      confirmationId,
+      confirmationNumber: String(confirmationNumber),
+      confirmationTimestamp,
+      isExecuted,
+      isConfirmed,
+      confirmations,
+      data: tx,
+    }
   }
 
 }
